View function for fetching non-expired allowlisted request digests

+++ contracts/src/v0.8/workflow/dev/v2/test/WorkflowRegistry/WorkflowRegistry.allowlistRequest.t.sol
@@ -8,7 +8,7 @@ contract WorkflowRegistry_allowlistRequest is WorkflowRegistrySetup {
   function test_allowlistRequest_WhenTheUserIsNotLinked() external {
     // it should revert with OwnershipLinkDoesNotExist
     bytes32 requestDigest = keccak256("request-digest");
-    uint256 expiryTimestamp = block.timestamp + 1 hours;
+    uint32 expiryTimestamp = uint32(block.timestamp + 1 hours);
 
     address vaultNode = address(0x89652);
     vm.prank(vaultNode);
@@ -19,13 +19,13 @@ contract WorkflowRegistry_allowlistRequest is WorkflowRegistrySetup {
     s_registry.allowlistRequest(requestDigest, expiryTimestamp);
 
     // old timestamp should revert
-    expiryTimestamp = block.timestamp - 1 hours;
+    expiryTimestamp = uint32(block.timestamp - 1 hours);
     vm.expectRevert(abi.encodeWithSelector(WorkflowRegistry.RequestExpired.selector, requestDigest, expiryTimestamp));
     vm.prank(s_user);
     s_registry.allowlistRequest(requestDigest, expiryTimestamp);
 
     // timestamp equal to current block timestamp should revert
-    expiryTimestamp = block.timestamp;
+    expiryTimestamp = uint32(block.timestamp);
     vm.expectRevert(abi.encodeWithSelector(WorkflowRegistry.RequestExpired.selector, requestDigest, expiryTimestamp));
     vm.prank(s_user);
     s_registry.allowlistRequest(requestDigest, expiryTimestamp);
@@ -34,7 +34,7 @@ contract WorkflowRegistry_allowlistRequest is WorkflowRegistrySetup {
   function test_allowlistRequest_WhenTheUserIsLinked() external {
     //it should allowlist the request digest
     bytes32 requestDigest = keccak256("request-digest");
-    uint256 expiryTimestamp = block.timestamp + 1 hours;
+    uint32 expiryTimestamp = uint32(block.timestamp + 1 hours);
 
     // link the owner first to ensure the request can be allowlisted
     _linkOwner(s_user);
@@ -51,7 +51,7 @@ contract WorkflowRegistry_allowlistRequest is WorkflowRegistrySetup {
     assertTrue(s_registry.isRequestAllowlisted(s_user, requestDigest), "Request should be allowlisted");
 
     bytes32 newRequestDigest = keccak256("new-request-digest");
-    uint256 newExpiryTimestamp = block.timestamp + 1 hours; // same timestamp as the previous request
+    uint32 newExpiryTimestamp = uint32(block.timestamp + 1 hours); // same timestamp as the previous request
     vm.expectEmit(true, true, true, false);
     emit WorkflowRegistry.RequestAllowlisted(s_user, newRequestDigest, newExpiryTimestamp);
     vm.prank(s_user);
@@ -66,7 +66,7 @@ contract WorkflowRegistry_allowlistRequest is WorkflowRegistrySetup {
     assertFalse(s_registry.isRequestAllowlisted(s_user, newRequestDigest), "New request should expire");
     assertFalse(s_registry.isRequestAllowlisted(s_user, requestDigest), "Old request should expire");
 
-    newExpiryTimestamp = block.timestamp + 2 hours; // same digest, but one hour ahead of block time
+    newExpiryTimestamp = uint32(block.timestamp + 2 hours); // same digest, but one hour ahead of block time
     vm.expectEmit(true, true, true, false);
     emit WorkflowRegistry.RequestAllowlisted(s_user, newRequestDigest, newExpiryTimestamp);
     vm.prank(s_user);
